added value storage

+++ edu.database.node/src/main/java/com/devtritus/edu/database/node/tree/BTreeNode.java
@@ -1,6 +1,6 @@
 package com.devtritus.edu.database.node.tree;
 
-class BTreeNode extends GenericBTreeNode<String, Integer, Integer> {
+class BTreeNode extends GenericBTreeNode<String, Long, Integer> {
     BTreeNode(int nodeId, int level) {
         super(nodeId, level, false);
     }
